refactor(create-assistant-ui): wrap assistant-ui create binary
Replace inline template selection and scaffolding logic with a
thin wrapper that locates and spawns the installed assistant-ui
binary, eliminating duplicated template definitions.
