+++ .changeset/rare-sheep-fly.md
@@ -1,0 +1,5 @@
+---
+"create-assistant-ui": patch
+---
+
+refactor(create-assistant-ui): delegate directly to the installed `assistant-ui create` binary instead of re-implementing template selection and scaffolding logic.
