+++ packages/create-assistant-ui/package.json
@@ -28,9 +28,7 @@
     "build": "aui-build"
   },
   "dependencies": {
-    "@clack/prompts": "^1.0.1",
-    "chalk": "^5.6.2",
-    "commander": "^14.0.3",
+    "assistant-ui": "workspace:*",
     "cross-spawn": "^7.0.6"
   },
   "devDependencies": {
